refactor: move and rename CurriculumDownloadButton
