+++ src/components/TeacherComponents/HeaderListingCurriculumDownloadButton/index.tsx
@@ -1,0 +1,1 @@
+export { default } from "./HeaderListingCurriculumDownloadButton";
